End of section 18.

+++ client/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -14,29 +15,37 @@ export class MessagesComponent implements OnInit {
   container='Unread';
   pageNumber = 1;
   pageSize = 5;
-  loading=false;
+ loading=false;
 
-  constructor(private messageService:MessageService) { }
+  constructor(private messageService:MessageService, private confirmService:ConfirmService) { }
   ngOnInit(): void {
     this.loadMessage()
   }
    loadMessage(){
-     this.loading=true;
-     this.messageService.getMessage(this.pageNumber, this.pageSize, this.container).subscribe(response=>{
+   this.loading=true;
+     this.messageService.getMessage(this.pageNumber, this.pageSize, this.container).subscribe(response=>{      
        this.messages= response.result;
        this.pagination = response.pagination;
-       this.loading=false;
+      this.loading=false;
      })
    }
 
    deleteMessage(id:number){
-     this.messageService.deleteMessage(id).subscribe(()=>{
-       this.messages.splice(this.messages.findIndex(m=>m.id === id),1)
-     })
+    
+    this.confirmService.confirm('Confirm delete messages','This cannot be undone!').subscribe(result =>{
+      if(result){        
+        this.messageService.deleteMessage(id).subscribe(()=>{
+          this.messages.splice(this.messages.findIndex(m=>m.id === id),1)
+        })
+      }
+    })
+
+  
    }
    pageChanged(evt:any){//problems with pagination
      this.pageNumber = evt.page;
      this.loadMessage();
-     //https://localhost:5001/api/messages?pageNumber=2&pageSize=4&Container=Inbox
+     //[hidden]="loading"
+    //&& !loading
    }
 }
